requirements.txt
# ==========================================
# 核心框架 (LangChain & LangGraph)
# ==========================================
langchain==0.3.18
langchain-core==0.3.34
langchain-openai==0.3.6
langchain-community==0.3.17
langchain-chroma==0.2.2
langchain-text-splitters==0.3.5
langchain-ollama==0.3.0
langgraph==0.3.18
langgraph-checkpoint==2.0.22

# ==========================================
# Web 框架
# ==========================================
fastapi==0.115.8
uvicorn==0.34.0
streamlit==1.42.2
python-multipart==0.0.20

# ==========================================
# 向量数据库 & 嵌入
# ==========================================
chromadb==0.5.23
sentence-transformers==3.4.1

# ==========================================
# 数据处理 & 配置
# ==========================================
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
pyyaml==6.0.2

# ==========================================
# 文件处理
# ==========================================
pypdf==5.3.0

# ==========================================
# 网络请求
# ==========================================
httpx==0.28.1
httpcore==1.0.7
requests==2.32.3

# ==========================================
# 工具类
# ==========================================
tenacity==9.0.0
tiktoken==0.9.0
numpy==1.26.4
